vizsurvey: Visualisation And Analysis During a Survey Field

vizsurvey is an R package designed to streamline the quality assessment of survey data by providing intuitive visual diagnostics through an interactive dashboard. vizsurvey is especially useful for institutions or researchers conducting large-scale surveys with multiple interviewers, enabling a fast and systematic overview of data quality over time.

See also

Author

Maintainer: Thomas Delclite thomas.delclite@economie.fgov.be

Authors:

Adrien Mierop adrien.mierop@economie.fgov.be